Restructure code and test end to end
