+++ Dockerfile
@@ -14,18 +14,21 @@ RUN chmod a+x /tmp/libssl1.1_1.1.1f-1ubuntu2_amd64.deb && \
 ARG GCP_CREDENTIALS
 ARG TOPIC_NAME 
 ARG SUBSCRIPTION_NAME
-ARG BASE_SERVER_URL
 ARG BITBUCKET_CLIENT_ID
 ARG BITBUCKET_CLIENT_SECRET
 ARG BITBUCKET_BASE_URL
+ARG INSTALL_ID
+ARG SERVER_URL
+
 
 ENV GCP_CREDENTIALS=$GCP_CREDENTIALS  
 ENV TOPIC_NAME=$TOPIC_NAME
 ENV SUBSCRIPTION_NAME=$SUBSCRIPTION_NAME
-ENV BASE_SERVER_URL=$BASE_SERVER_URL
 ENV BITBUCKET_CLIENT_ID=$BITBUCKET_CLIENT_ID
 ENV BITBUCKET_CLIENT_SECRET=$BITBUCKET_CLIENT_SECRET
 ENV BITBUCKET_BASE_URL=$BITBUCKET_BASE_URL
+ENV INSTALL_ID=$INSTALL_ID
+ENV SERVER_URL=$SERVER_URL
 
 COPY ./devprofiler/target/debug/devprofiler /app/devprofiler
 COPY ./pubsub-sa.json /app/pubsub-sa.json
