+++ devprofiler/Cargo.toml
@@ -33,5 +33,6 @@ uuid = { version = "1.4.0", features = ["v4"]}
 rand = "0.8.5"
 dotenv = "0.15.0"
 chrono = "0.4.26"
+tonic = "0.9.2"
 
 # todo - check all lib licences
