+++ devprofiler/src/pubsub/mod.rs
@@ -1,2 +1,1 @@
 pub mod listener;
-pub mod publisher;
